+++ MyVocaList/MauiProgram.cs
@@ -1,6 +1,9 @@
-﻿using CommunityToolkit.Maui;
+using CommunityToolkit.Maui;
 using Microsoft.Extensions.Logging;
 using UraniumUI;
+#if ANDROID || IOS || MACCATALYST
+using HorusStudio.Maui.MaterialDesignControls;
+#endif
 
 namespace MyVocaList;
 
@@ -12,15 +15,23 @@ public static class MauiProgram
 		builder
 			.UseMauiApp<App>()
 			.UseMauiCommunityToolkit()
-			//.ConfigureMopups()
 			.UseUraniumUI()
 			.UseUraniumUIMaterial()
+#if ANDROID || IOS || MACCATALYST
+			.UseMaterialDesignControls(ConfigureMDC)
+#endif
 			.ConfigureFonts(fonts =>
 			{
-				fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
-				fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
+				// MD3 Default: Roboto
+				fonts.AddFont("Roboto-Regular.ttf", "RobotoRegular");
+				fonts.AddFont("Roboto-Medium.ttf", "RobotoMedium");
+				fonts.AddFont("Roboto-Bold.ttf", "RobotoBold");
+
+				// Material Symbols Icons
+				fonts.AddMaterialSymbolsFonts();
 			});
-			
+
+		// Register Services (ThreadSafeDialogService added in Task 5)
 
 #if DEBUG
 		builder.Logging.AddDebug();
@@ -28,4 +39,24 @@ public static class MauiProgram
 
 		return builder.Build();
 	}
+
+#if ANDROID || IOS || MACCATALYST
+	private static void ConfigureMDC(MaterialDesignControlsBuilder options)
+	{
+#if DEBUG
+		options.EnableDebug();
+#endif
+		options.OnException((sender, ex) =>
+		{
+			System.Diagnostics.Debug.WriteLine($"[MDC] {sender}: {ex}");
+		});
+
+		options.ConfigureThemesFromResources();
+		options.ConfigureStringFormat(new MaterialFormatOptions
+		{
+			DateFormat = "dd/MM/yyyy",
+			TimeFormat = "HH:mm"
+		});
+	}
+#endif
 }